add the endpoint to create and update faq items
