+++ src/main/java/zas/admin/zec/backend/config/SecurityConfig.java
@@ -35,10 +35,11 @@ public class SecurityConfig {
                 .sessionManagement(smc -> smc.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers(HttpMethod.POST,"/api/users").permitAll()
-                        .requestMatchers(HttpMethod.GET, "/api/faqs/**").permitAll()
-                        .requestMatchers("/api/admin/**").hasAnyAuthority("ADMIN")
+                        .requestMatchers(HttpMethod.GET, "/api/faq-items/**").permitAll()
+                        .requestMatchers("/api/admin/**", "/api/faq-items/**").hasAnyAuthority("ADMIN")
                         .anyRequest().authenticated())
                 .httpBasic(Customizer.withDefaults())
+                .formLogin(AbstractHttpConfigurer::disable)
                 .authenticationProvider(daoAuthenticationProvider());
 
         return http.build();
