+++ src/main/java/zas/admin/zec/backend/faqs/FAQItemsController.java
@@ -1,0 +1,42 @@
+package zas.admin.zec.backend.faqs;
+
+import jakarta.annotation.Nullable;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@RestController
+@RequestMapping("/api/faq-items")
+public class FAQItemsController {
+
+    public record FAQItem(@Nullable Long id, String text, String answer, String url, String language) {}
+    @Qualifier("pyBackendWebClient")
+    private final WebClient pyBackendWebClient;
+
+    public FAQItemsController(WebClient pyBackendWebClient) {
+        this.pyBackendWebClient = pyBackendWebClient;
+    }
+
+    @GetMapping
+    public ResponseEntity<Object> searchFAQItems(@RequestParam String question) {
+        return pyBackendWebClient.get()
+                .uri(uriBuilder -> uriBuilder.path("/apy/autocomplete/")
+                        .queryParam("question", question)
+                        .build())
+                .retrieve()
+                .toEntity(Object.class)
+                .block();
+    }
+
+    @PutMapping
+    public ResponseEntity<Object> updateFAQItem(@RequestBody FAQItem faqItem) {
+        return pyBackendWebClient.put()
+                .uri(uriBuilder -> uriBuilder.path("/apy/indexing/data")
+                        .build())
+                .bodyValue(faqItem)
+                .retrieve()
+                .toEntity(Object.class)
+                .block();
+    }
+}
